+++ ApocalipseCHI/src/Objeto.java
@@ -18,6 +18,6 @@ public abstract class Objeto {
     
     abstract public void DesenhaSe(Graphics2D dbg,int Xmundo,int Ymundo);
 
-    abstract public void SimulaSe(long diftime);    
+    abstract public void SimulaSe(float diftime);    
     
 }
